+++ downstream-test/src/main/java/arez/downstream/BuildDownstream.java
@@ -29,7 +29,7 @@ public final class BuildDownstream
   private static void run()
     throws Exception
   {
-    Gir.go( () -> Stream.of( "arez-dom", "arez-promise", "arez-spytools" )
+    Gir.go( () -> Stream.of( "arez-dom", "arez-promise", "arez-spytools", " arez-testng" )
       .forEach( name -> WorkspaceUtil.forEachBranch( name,
                                                      "https://github.com/arez/" + name + ".git",
                                                      Collections.singletonList( "master" ),