+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.57](https://github.com/arez/arez-dom/tree/v0.57) (2020-02-21) · [Full Changelog](https://github.com/arez/arez-dom/compare/v0.56...v0.57)
 
 * Update the 'org.realityforge.arez' dependencies to version '0.170'