+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.01](https://github.com/arez/arez-testng/tree/v0.01) (2020-03-09) · [Full Changelog](https://github.com/arez/arez-testng/compare/v0.00...v0.01)
 
  ‎🎉	Initial release ‎🎉