+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Revert the `ArezTestSupport` methods `beforeMethod()` and `afterMethod()` back to `preTest()` and `postTest()` as there is just too much existing code that uses the old convention.
+
 ### [v0.01](https://github.com/arez/arez-testng/tree/v0.01) (2020-03-09) · [Full Changelog](https://github.com/arez/arez-testng/compare/v0.00...v0.01)
 
  ‎🎉	Initial release ‎🎉