+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.02](https://github.com/arez/arez-testng/tree/v0.02) (2020-03-09) · [Full Changelog](https://github.com/arez/arez-testng/compare/v0.01...v0.02)
 
 * Revert the `ArezTestSupport` methods `beforeMethod()` and `afterMethod()` back to `preTest()` and `postTest()` as there is just too much existing code that uses the old convention.