+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.26.0`.
 * Add integration tests to verify sting integration behaves as expected.
 * Add the `arez-testng` project to the set of related projects and the set of downstream projects that are tested as part of the release process.
+* Ensure that invoking `ArezTestUtil.enableSpies()` and `ArezTestUtil.disableSpies()` resets arez state. If this is not done it can leave Arez in an inconsistent state during testing. i.e. A spy object not present when it is expected to be present.
 
 ### [v0.170](https://github.com/arez/arez/tree/v0.170) (2020-02-21) · [Full Changelog](https://github.com/arez/arez/compare/v0.169...v0.170)